requirements.txt
discord.py
graia-application-mirai
websockets